+++ libiap/__init__.py
@@ -1,4 +1,4 @@
-__version__ = VERSION = '0.2.0rc1'
+__version__ = VERSION = '0.2.0'
 
 import logging
 logging.getLogger(__name__).addHandler(logging.NullHandler())