+++ libiap/openapi/libconsole/api/__init__.py
@@ -1,0 +1,11 @@
+from __future__ import absolute_import
+
+# flake8: noqa
+
+# import apis into api package
+from libiap.openapi.libconsole.api.accounts_api import AccountsApi
+from libiap.openapi.libconsole.api.health_api import HealthApi
+from libiap.openapi.libconsole.api.regions_api import RegionsApi
+from libiap.openapi.libconsole.api.tokens_api import TokensApi
+from libiap.openapi.libconsole.api.usages_api import UsagesApi
+from libiap.openapi.libconsole.api.workgroups_api import WorkgroupsApi