+++ libiap/openapi/libgds/api/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+
+# flake8: noqa
+
+# import apis into api package
+from libiap.openapi.libgds.api.files_api import FilesApi
+from libiap.openapi.libgds.api.folders_api import FoldersApi
+from libiap.openapi.libgds.api.volumes_api import VolumesApi