+++ libiap/openapi/libwes/api/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import
+
+# flake8: noqa
+
+# import apis into api package
+from libiap.openapi.libwes.api.workflow_runs_api import WorkflowRunsApi
+from libiap.openapi.libwes.api.workflow_signals_api import WorkflowSignalsApi
+from libiap.openapi.libwes.api.workflow_versions_api import WorkflowVersionsApi
+from libiap.openapi.libwes.api.workflows_api import WorkflowsApi