+++ CHANGELOG.rst
@@ -2,3 +2,13 @@
 -----
 * Release date: 2020-05-01
 * Initial release with implementation to all GET API calls
+
+0.2.0
+-----
+* Release date: 2020-06-12
+* Added openapi-generator[1] auto generated SDK
+* Added Prism[2] for integration test mock server
+* Added ``syncapi.sh`` script and ``Makefile`` workflow for repeatable auto SDK generation
+
+#. [1]: https://github.com/OpenAPITools/openapi-generator
+#. [2]: https://stoplight.io/open-source/prism/